requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
airsim==1.2.8
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cloudpickle==1.3.0
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
future==0.18.2
gast==0.3.3
google-auth==1.20.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
gym==0.17.2
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
joblib==1.0.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
mathutils==2.81.2
matplotlib==3.3.1
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
multidict==4.7.6
networkx==2.4
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.40
opencv-python==4.4.0.42
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyglet==1.5.0
pynput==1.6.8
pyparsing==2.4.7
pyquaternion==0.9.5
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.24.0
scipy==1.5.2
six==1.15.0
tabulate==0.8.7
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.8.13
torch==1.6.0
torchvision==0.7.0
tornado==4.5.3
tqdm==4.48.2
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
